C - Binary Trees. Task 15. Is full

+++ 14-binary_tree_balance.c
@@ -8,9 +8,9 @@
  * Return: the height of the tree
  *
  */
-size_t binary_tree_height(const binary_tree_t *tree)
+int binary_tree_height(const binary_tree_t *tree)
 {
-	size_t height_l = 0, height_r = 0;
+	int height_l = 0, height_r = 0;
 
 	if ((tree == NULL) || ((tree->left == NULL) && (tree->right == NULL)))
 		return (0);
@@ -34,7 +34,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t height_l = 0, height_r = 0;
+	int height_l = 0, height_r = 0;
 
 	if ((tree == NULL) || ((tree->left == NULL) && (tree->right == NULL)))
 		return (0);
